Fixes from adam-parker1

+++ src/map_types/dim_mapping.cpp
@@ -17,5 +17,5 @@ libtokamap::TypedDataArray libtokamap::DimMapping::map(const MapArguments& argum
         throw libtokamap::MappingError{"cannot use DIM_PROBE on rank 0 mapping '" + m_dim_probe + "'"};
     }
     // TODO: Add DIM_INDEX for selecting which dimension to return size of
-    return TypedDataArray{array.shape()[0]};
+    return TypedDataArray{static_cast<int>(array.shape()[0])};
 }
